requirements.txt
# core
ibis-framework
cached_property
google-cloud-bigquery
pydata-google-auth

# dev
black==22.3.0
pytest
pytest-cov
pytest-mock
mock
flake8
flake8-comprehensions
mypy
isort
pydocstyle
setuptools-scm
types-pytz
types-pkg_resources
types-setuptools
